Better buffer management, cleaned code, improved gaussian blur, reworked bloom,...

+++ shaders/bufferSettings.glsl
@@ -22,7 +22,7 @@ const int colortex3Format = RGBA16F;
 const bool colortex3Clear = false;
 
 MULTIPLE THINGS
-const int colortex4Format = RGBA16F;
+const int colortex4Format = RGBA16;
 const int colortex5Format = RGBA16F;
 
 PTGI ACCUMULATION BUFFER
